requirements.txt
Django==1.11.6
djangorestframework